feat : output을 delegate를 통해 전달할 수 있도록 MeasureViewDelegate 구현
- 좀 더 명시적으로 MeasureDelegate -> MeasureServiceDelegate 로 네이밍 수정
